feat: replace lspinstaller with mason
since it acts as a general package manager for editor related stuff like
debuggers, language servers and is not restricted to lsp only.

first step into making neovim completely independent of the system
(right now we rely on the system has language servers installed. it was
a decision that we kind of regret since it makes sense for the editor to
be self contained and not rely on the multiple systems you might develop
in)